+++ SWFModernRuntime/include/libswf/tag.h
@@ -72,11 +72,11 @@ void tagSetClipActions(SWFAppContext* app_context, size_t depth, ClipAction* cli
 void tagPlaceObject3(SWFAppContext* app_context, size_t depth, size_t char_id,
     u32 transform_id, u32 cxform_id, u16 clip_depth, u8 blend_mode);
 void tagSetFilter(SWFAppContext* app_context, size_t depth,
-    u8 type, float blur_x, float blur_y, u8 quality, u8 flags,
-    float r, float g, float b, float a, float strength,
-    float angle, float distance);
+    u8 type, double blur_x, double blur_y, u8 quality, u8 flags,
+    double r, double g, double b, double a, double strength,
+    double angle, double distance);
 void tagSetFilterHighlight(SWFAppContext* app_context, size_t depth,
-    float r, float g, float b, float a);
+    double r, double g, double b, double a);
 void tagSetInstanceName(SWFAppContext* app_context, size_t depth, const char* name);
 void tagRemoveObject(SWFAppContext* app_context, size_t depth);
 void tagRemoveObject2(SWFAppContext* app_context, size_t depth);
